Raycasting template

+++ src/structures.h
@@ -20,18 +20,19 @@ public:
 	bool green;
 	bool blue;
 
+	// Converts to sf::Color object
 	sf::Color convertToColor();
 };
 
 class Ray
 {
 public:
-	Ray(float xx_start, float yy_start, float xx_end, float yy_end, Color col) : x_start(xx_start), y_start(yy_start), x_end(xx_end), y_end(yy_end), color(col) {}
+	Ray(double xx_start, double yy_start, double xx_end, double yy_end, Color col) : x_start(xx_start), y_start(yy_start), x_end(xx_end), y_end(yy_end), color(col) {}
 
-	float x_start;
-	float y_start;
-	float x_end;
-	float y_end;
+	double x_start;
+	double y_start;
+	double x_end;
+	double y_end;
 	Color color;
 };
